requirements.txt
fastapi
uvicorn
torch
transformers
PyGithub
pandas
requests